requirements.txt
django
django-cleanup
django-constance
django-import-export
gunicorn
pillow
python-dotenv
